feat: 유저 게시글 맵핑 추가

+++ src/main/java/org/example/newsfeed/like/service/LikeServiceImpl.java
@@ -1,12 +1,17 @@
 package org.example.newsfeed.like.service;
 
 import lombok.RequiredArgsConstructor;
+import org.example.newsfeed.global.exception.BaseException;
 import org.example.newsfeed.like.dto.LikeResponseDto;
 import org.example.newsfeed.like.entity.Like;
 import org.example.newsfeed.like.repository.LikeRepository;
+import org.example.newsfeed.post.entitiy.Post;
+import org.example.newsfeed.post.repository.PostRepository;
+import org.example.newsfeed.user.entity.User;
+import org.example.newsfeed.user.repository.UserRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -14,12 +19,18 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class LikeServiceImpl implements LikeService {
     private final LikeRepository likeRepository;
+    private final UserRepository userRepository;
+    private final PostRepository postRepository;
 
     @Override
-    public LikeResponseDto createLike(Long postId) {
-        //todo : 추후 UserId JWT를 이용하여 맵핑
-        Long userId = 1L;
-        Like like = new Like(userId, postId);
+    public LikeResponseDto createLike(Long postId, Long userId) {
+
+        // 유저데이터 조회
+        User user = userRepository.findById(userId).orElseThrow();
+        // 게시글 조회
+        Post post = postRepository.findById(postId).orElseThrow();
+
+        Like like = new Like(user, post);
         likeRepository.save(like);
 
         return new LikeResponseDto(like);
@@ -34,8 +45,15 @@ public class LikeServiceImpl implements LikeService {
     }
 
     @Override
-    public void deleteLike(Long likeId) {
+    public void deleteLike(Long likeId, Long userId) {
+        // 유저데이터 조회
+        User user = userRepository.findById(userId).orElseThrow();
+
         Like like = likeRepository.findById(likeId).orElseThrow();
+
+        if (!user.getId().equals(like.getUser().getId())) {
+            throw new BaseException(HttpStatus.FORBIDDEN, "권한이 없습니다");
+        }
 
         likeRepository.delete(like);
     }
